+++ formatters/hyperlinks.js
@@ -8,7 +8,7 @@ const hyperlinks = require('../lib/hyperlinks');
  */
 function generateHyperlinkReference (hyperlink = '') {
   /** Check the URL */
-  hyperlink = hyperlinks.validateURL(hyperlink, 'docx');
+  hyperlink = hyperlinks.validateURL(hyperlink, this.defaultUrlOnError, 'docx');
   hyperlinks.addLinkDatabase(this, hyperlink);
   return {
     fn   : generateHyperkinReferencePostProcessing,
@@ -38,10 +38,25 @@ function generateHyperkinReferencePostProcessing (hyperlink) {
  * @param {String} hyperlink
  */
 function validateURL (hyperlink = '') {
-  return hyperlinks.validateURL(hyperlink);
+  return hyperlinks.validateURL(hyperlink, this.defaultUrlOnError);
+}
+
+/**
+ * Replace the default value hyperlinks.URL_ON_ERROR when dynamic hyperlinks are injected into a report.
+ * If the `newDefaultUrlOnError` argument is invalid, it is replaced by the default hyperlinks.URL_ON_ERROR.
+ *
+ * @private
+ * @param {String} content Main content coming from the marker.
+ * @param {String} newDefaultUrlOnError new URL replacing the hyperlinks.URL_ON_ERROR
+ * @returns {String} the content
+ */
+function defaultURL (content, newDefaultUrlOnError = '') {
+  this.defaultUrlOnError = hyperlinks.validateURL(newDefaultUrlOnError);
+  return content;
 }
 
 module.exports = {
   generateHyperlinkReference,
-  validateURL
+  validateURL,
+  defaultURL
 };